[HACK] proto_holtek: fix writeFirmware operation
Hacks a new control transfer operation to hid and hiddevice so that it
can be used to wait for the completion of individual write operations in
proto_holtek.

This targets Linux with LibUSB-1.0 for now. Plan is to support macOS and
Windows as LibUSB-1.0 becomes standard backend.

With this, writeFirmware is confirmed to flash properly albeit at a less
optimal pace. Future plan involves writing larger chunks then waiting
for the chunk to be written before moving to the next chunk.